refactor: refactor api for user
