+++ backend/test/users.test.ts
@@ -621,54 +621,6 @@ describe('users api', () => {
     expect(payload.every((row) => typeof row.color === 'string' && row.color.length > 0)).toBe(true);
   });
 
-  it('does not append admins for staff when includeAdmins=1', async () => {
-    state.jwtSub = 'auth0|staff-user';
-
-    const response = await app.request('/api/users/room-1?includeAdmins=1', {
-      method: 'GET',
-    }, env);
-
-    expect(response.status).toBe(200);
-    const payload = (await response.json()) as Array<{ id: string; role: string }>;
-    const ids = payload.map((row) => row.id).sort();
-    expect(ids).toEqual(['auth0|manager-user', 'auth0|staff-user'].sort());
-    expect(ids).not.toContain('auth0|admin-user');
-    expect(payload.every((row) => !('email' in row))).toBe(true);
-  });
-
-  it('appends admins when includeAdmins=1 for manager', async () => {
-    state.jwtSub = 'auth0|manager-user';
-
-    const response = await app.request('/api/users/room-1?includeAdmins=1', {
-      method: 'GET',
-    }, env);
-
-    expect(response.status).toBe(200);
-    const payload = (await response.json()) as Array<{ id: string; role: string }>;
-    const ids = payload.map((row) => row.id);
-    expect(ids).toContain('auth0|staff-user');
-    expect(ids).toContain('auth0|admin-user');
-    expect(ids).toContain('auth0|other-admin');
-    expect(payload.find((r) => r.id === 'auth0|admin-user')?.role).toBe('admin');
-    expect(payload.every((row) => !('email' in row))).toBe(true);
-  });
-
-  it('appends admins with emails when includeAdmins=1 for admin', async () => {
-    state.jwtSub = 'auth0|admin-user';
-
-    const response = await app.request('/api/users/room-1?includeAdmins=1', {
-      method: 'GET',
-    }, env);
-
-    expect(response.status).toBe(200);
-    const payload = (await response.json()) as Array<{ id: string; role: string; email?: string }>;
-    const ids = payload.map((row) => row.id);
-    expect(ids).toContain('auth0|staff-user');
-    expect(ids).toContain('auth0|admin-user');
-    expect(ids).toContain('auth0|other-admin');
-    expect(payload.find((r) => r.id === 'auth0|staff-user')?.email).toBe('staff@example.com');
-  });
-
   it('returns 403 when manager requests another classroom users', async () => {
     state.jwtSub = 'auth0|manager-user';
