+++ shared/type.ts
@@ -1,0 +1,12 @@
+/**
+ * 型など
+ */
+export type User = {
+  id: string;
+  role: 'admin' | 'manager' | 'staff';
+  classroomId: string | null;
+  firstName: string;
+  lastName: string;
+  email: string;
+  color: string;
+};
