+++ tsconfig.worker.json
@@ -4,5 +4,5 @@
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.worker.tsbuildinfo",
     "types": ["./worker-configuration.d.ts"],
   },
-  "include": ["backend/**/*", "worker-configuration.d.ts"]
+  "include": ["backend/**/*", "worker-configuration.d.ts", "shared/type.ts"]
 }
